% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runApp}
\alias{runApp}
\title{Run shiny application}
\usage{
runApp(launch_browser = FALSE, port = 8080)
}
\arguments{
\item{launch_browser}{...}

\item{port}{...}
}
\description{
This is a master function used to start an application implemented by this
package.
}
