% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{shinyServer}
\alias{shinyServer}
\title{<ToDo>}
\usage{
shinyServer(input, output, session)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
<ToDo>
}
\examples{
add(1, 1)
add(10, 1)

}
