% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runApp}
\alias{runApp}
\title{<ToDo>}
\usage{
runApp()
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
<ToDo>
}
\examples{
add(1, 1)
add(10, 1)

}
